Adds descriptor set for textures (NOT WORKING)
The program will not compile at the moment, because the time when the graphics pipeline gets created does not fit into the new descriptor set scheme. Right now, the graphics pipeline is created before the scene is filled with data. However, the graphics pipeline has to know how many textures there are to get created. So a new order has to be figured out.